Reduce more code
