+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/common/tagdetails/AbstractTagToken.java
@@ -13,10 +13,17 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+
+import org.eclipse.hawkbit.repository.SpPermissionChecker;
 import org.eclipse.hawkbit.ui.management.state.ManagementUIState;
+import org.eclipse.hawkbit.ui.utils.I18N;
 import org.eclipse.hawkbit.ui.utils.SPUIDefinitions;
 import org.eclipse.hawkbit.ui.utils.SPUIStyleDefinitions;
+import org.eclipse.hawkbit.ui.utils.UINotification;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.vaadin.spring.events.EventBus;
 import org.vaadin.tokenfield.TokenField;
 import org.vaadin.tokenfield.TokenField.InsertPosition;
 
@@ -52,13 +59,33 @@ public abstract class AbstractTagToken implements Serializable {
     protected final Map<Long, TagData> tokensAdded = new HashMap<>();
 
     protected CssLayout tokenLayout = new CssLayout();
+    
+    @Autowired
+    protected SpPermissionChecker checker;
 
+    @Autowired
+    protected I18N i18n;
+
+    @Autowired
+    protected UINotification uinotification;
+
+    @Autowired
+    protected transient EventBus.SessionEventBus eventBus;
+    
     @Autowired
     protected ManagementUIState managementUIState;
 
+    @PostConstruct
     protected void init() {
         createTokenField();
         checkIfTagAssignedIsAllowed();
+        eventBus.subscribe(this);
+    }
+   
+
+    @PreDestroy
+    void destroy() {
+        eventBus.unsubscribe(this);
     }
 
     private void createTokenField() {
