+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/common/tagdetails/DistributionTagToken.java
@@ -14,7 +14,6 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 
 import org.eclipse.hawkbit.eventbus.event.DistributionSetTagAssigmentResultEvent;
@@ -22,7 +21,6 @@ import org.eclipse.hawkbit.eventbus.event.DistributionSetTagCreatedBulkEvent;
 import org.eclipse.hawkbit.eventbus.event.DistributionSetTagDeletedEvent;
 import org.eclipse.hawkbit.eventbus.event.DistributionSetTagUpdateEvent;
 import org.eclipse.hawkbit.repository.DistributionSetManagement;
-import org.eclipse.hawkbit.repository.SpPermissionChecker;
 import org.eclipse.hawkbit.repository.TagManagement;
 import org.eclipse.hawkbit.repository.model.DistributionSet;
 import org.eclipse.hawkbit.repository.model.DistributionSetTag;
@@ -31,10 +29,7 @@ import org.eclipse.hawkbit.ui.management.event.DistributionTableEvent;
 import org.eclipse.hawkbit.ui.management.event.DistributionTableEvent.DistributionComponentEvent;
 import org.eclipse.hawkbit.ui.management.event.ManagementUIEvent;
 import org.eclipse.hawkbit.ui.utils.HawkbitCommonUtil;
-import org.eclipse.hawkbit.ui.utils.I18N;
-import org.eclipse.hawkbit.ui.utils.UINotification;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.vaadin.spring.events.EventBus;
 import org.vaadin.spring.events.EventScope;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 
@@ -52,19 +47,6 @@ import com.vaadin.ui.UI;
 public class DistributionTagToken extends AbstractTagToken {
 
     private static final long serialVersionUID = -8022738301736043396L;
-
-    @Autowired
-    private SpPermissionChecker spChecker;
-
-    @Autowired
-    private I18N i18n;
-
-    @Autowired
-    private UINotification uinotification;
-
-    @Autowired
-    private transient EventBus.SessionEventBus eventBus;
-
     @Autowired
     private transient TagManagement tagManagement;
 
@@ -73,19 +55,8 @@ public class DistributionTagToken extends AbstractTagToken {
 
     private DistributionSet selectedDS;
 
-    private UI ui;
-
     // To Be Done : have to set this value based on view???
     private static final Boolean NOTAGS_SELECTED = Boolean.FALSE;
-
-    @Override
-    @PostConstruct
-    protected void init() {
-        super.init();
-        ui = UI.getCurrent();
-        eventBus.subscribe(this);
-
-    }
 
     @Override
     protected String getTagStyleName() {
@@ -140,7 +111,7 @@ public class DistributionTagToken extends AbstractTagToken {
 
     @Override
     protected Boolean isToggleTagAssignmentAllowed() {
-        return spChecker.hasUpdateDistributionPermission();
+        return checker.hasUpdateDistributionPermission();
     }
 
     @Override
@@ -163,18 +134,15 @@ public class DistributionTagToken extends AbstractTagToken {
 
     @EventBusListenerMethod(scope = EventScope.SESSION)
     void onEvent(final DistributionTableEvent distributionTableEvent) {
-        if (distributionTableEvent.getDistributionComponentEvent() == DistributionComponentEvent.ON_VALUE_CHANGE) {
-            ui.access(() -> {
-                /**
-                 * distributionTableEvent.getDistributionSet() is null when
-                 * table has no data.
-                 */
-                if (distributionTableEvent.getDistributionSet() != null) {
-                    selectedDS = distributionTableEvent.getDistributionSet();
-                    repopulateToken();
-                }
-            });
+        if (distributionTableEvent.getDistributionComponentEvent() != DistributionComponentEvent.ON_VALUE_CHANGE) {
+            return;
         }
+        UI.getCurrent().access(() -> {
+            if (distributionTableEvent.getDistributionSet() != null) {
+                selectedDS = distributionTableEvent.getDistributionSet();
+                repopulateToken();
+            }
+        });
     }
 
     @EventBusListenerMethod(scope = EventScope.SESSION)
@@ -234,6 +202,7 @@ public class DistributionTagToken extends AbstractTagToken {
         return false;
     }
 
+    @Override
     @PreDestroy
     void destroy() {
         eventBus.unsubscribe(this);
